// src/services/meta-data-scrapper/interfaces/property-features.interface.ts
export interface PropertyFeatures {
  duplex: boolean | null;
  triplex: boolean | null;
  loft: boolean | null;

  terrasse: boolean | null;
  balcon: boolean | null;
  loggia: boolean | null;
  jardin: boolean | null;
  patio: boolean | null;

  piscine: boolean | null;
  jacuzzi: boolean | null;
  spa: boolean | null;
  sauna: boolean | null;

  parking: boolean | null;
  garage: boolean | null;
  box: boolean | null;
  cave: boolean | null;
  grenier: boolean | null;

  ascenseur: boolean | null;
  gardien: boolean | null;
  digicode: boolean | null;
  interphone: boolean | null;
  visiophone: boolean | null;

  climatisation: boolean | null;
  cheminee: boolean | null;
  cuisineEquipee: boolean | null;
  dressing: boolean | null;
  buanderie: boolean | null;

  vueMer: boolean | null;
  vueMontagne: boolean | null;
  vuePanoramique: boolean | null;
  vueDegagee: boolean | null;

  dernierEtage: boolean | null;
  traversant: boolean | null;
  lumineux: boolean | null;
  calme: boolean | null;
  renove: boolean | null;
  standing: boolean | null;
  prestige: boolean | null;
}